* EZGraph.m: Don't include collections.h.  Include simtoolsgui.h. Reformatting throughout.
* EZDistribution.m ([EZDistribution -createEnd]): Include misc.h.
Use xmalloc instead of malloc.

* EZBin.m: Don't include collections.h or gui.h.  Include misc.h.
([EZBin -createEnd]): Use xmalloc for distribution, cachedLimits,
and locations.

* EZBin.h: Include simtoolsgui/GUIComposite.h, not
simtools/GUIComposite.h.

* Averager.h: Reformatting throughout.


git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@638 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/analysis/Averager.m
@@ -72,10 +72,10 @@
 
 - (double)getAverage
 {
-  if(count)
-    return total/count;
+  if (count)
+    return total / (double)count;
   else 
-    return 0;
+    return 0.0;
 } 
 
 - (double)getTotal
